// lib/database.types.ts
export type Database = {
// ...
  public: {
    Tables: {
      comment: {
        Row: {
          created_at: string | null
          dislikeCount: number | null
          id: string
          likeCount: number | null
          text: string | null
          user_id: string | null
          video_id: string | null
        }
        Insert: {
          created_at?: string | null
          dislikeCount?: number | null
          id?: string
          likeCount?: number | null
          text?: string | null
          user_id?: string | null
          video_id?: string | null
        }
        Update: {
          created_at?: string | null
          dislikeCount?: number | null
          id?: string
          likeCount?: number | null
          text?: string | null
          user_id?: string | null
          video_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "comment_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "comment_video_id_fkey"
            columns: ["video_id"]
            isOneToOne: false
            referencedRelation: "video"
            referencedColumns: ["id"]
          },
        ]
      }
      likedComments: {
        Row: {
          comment_id: string | null
          created_at: string | null
          id: string
          like: boolean | null
          user_id: string | null
        }
        Insert: {
          comment_id?: string | null
          created_at?: string | null
          id?: string
          like?: boolean | null
          user_id?: string | null
        }
        Update: {
          comment_id?: string | null
          created_at?: string | null
          id?: string
          like?: boolean | null
          user_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "likedComments_comment_id_fkey"
            columns: ["comment_id"]
            isOneToOne: false
            referencedRelation: "comment"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "likedComments_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
      likedVideos: {
        Row: {
          created_at: string | null
          id: string
          liked: boolean | null
          user_id: string | null
          video_id: string | null
        }
        Insert: {
          created_at?: string | null
          id?: string
          liked?: boolean | null
          user_id?: string | null
          video_id?: string | null
        }
        Update: {
          created_at?: string | null
          id?: string
          liked?: boolean | null
          user_id?: string | null
          video_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "likedVideos_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "likedVideos_video_id_fkey"
            columns: ["video_id"]
            isOneToOne: false
            referencedRelation: "video"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      profiles: {
        Row: {
          avatar_url: string | null
          full_name: string | null
          id: string
          updated_at: string | null
          username: string | null
        }
        Insert: {
          avatar_url?: string | null
          full_name?: string | null
          id?: string
          updated_at?: string | null
          username?: string | null
        }
        Update: {
          avatar_url?: string | null
          full_name?: string | null
          id?: string
          updated_at?: string | null
          username?: string | null
        }
        Relationships: []
      }
// ...
      video: {
        Row: {
          created_at: string | null
          description: string | null
          dislikes: number | null
          id: string
          likes: number | null
          thumbnailUrl: string | null
          title: string | null
          user_id: string | null
          videoStatus: boolean | null
          videoUrl: string | null
          viewCount: number | null
        }
        Insert: {
          created_at?: string | null
          description?: string | null
          dislikes?: number | null
          id?: string
          likes?: number | null
          thumbnailUrl?: string | null
          title?: string | null
          user_id?: string | null
          videoStatus?: boolean | null
          videoUrl?: string | null
          viewCount?: number | null
        }
        Update: {
          created_at?: string | null
          description?: string | null
          dislikes?: number | null
          id?: string
          likes?: number | null
          thumbnailUrl?: string | null
          title?: string | null
          user_id?: string | null
          videoStatus?: boolean | null
          videoUrl?: string | null
          viewCount?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "video_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}